feat: refactor PokémonDetails component to utilize modular subcomponents for better organization and maintainability; implement species details and evolution chain display
