+++ src/pages/PokemonDetails/components/PokemonSpritesCard.tsx
@@ -1,0 +1,39 @@
+import { Card, Image } from "antd";
+import type { Pokemon } from "../../../store/pokemon/types";
+
+type Props = {
+  pokemon: Pokemon;
+};
+
+export function PokemonSpritesCard({ pokemon }: Props) {
+  return (
+    <Card className="pokemon-info-card" title="Sprites" style={{ marginTop: 16 }}>
+      <div className="pokemon-sprites">
+        {!pokemon.sprites.front_default && !pokemon.sprites.back_default ? (
+          <span>No sprites available</span>
+        ) : (
+          <>
+            {pokemon.sprites.front_default && (
+              <Image
+                src={pokemon.sprites.front_default}
+                alt={`${pokemon.name} front`}
+                className="sprite"
+                draggable={false}
+                preview={false}
+              />
+            )}
+            {pokemon.sprites.back_default && (
+              <Image
+                src={pokemon.sprites.back_default}
+                alt={`${pokemon.name} back`}
+                className="sprite"
+                draggable={false}
+                preview={false}
+              />
+            )}
+          </>
+        )}
+      </div>
+    </Card>
+  );
+}
