+++ src/pages/PokemonDetails/components/index.ts
@@ -1,0 +1,7 @@
+export { PokemonHeroCard } from "./PokemonHeroCard";
+export { PokemonTypesCard } from "./PokemonTypesCard";
+export { PokemonSpritesCard } from "./PokemonSpritesCard";
+export { PokemonStatsCard } from "./PokemonStatsCard";
+export { PokemonAboutCard } from "./PokemonAboutCard";
+export type { SpeciesDetails } from "./PokemonAboutCard";
+export { PokemonEvolutionCard } from "./PokemonEvolutionCard";
